+++ raphodo/utilities.py
@@ -40,6 +40,7 @@ import struct
 import ctypes
 import signal
 import warnings
+import xdg
 import babel
 from glob import glob
 from pkg_resources import parse_version
@@ -855,6 +856,78 @@ def extract_file_from_tar(full_tar_path, member_filename) -> bool:
         except OSError:
             logging.error('Unable to move %s to new location', member_filename)
             return False
+
+
+def bug_report_full_tar_path() -> str:
+    """
+    Generate a full path for uncompressed bug report tar file.
+    The filename will not already exist.
+
+    :return: File name including path
+    """
+
+    filename = 'rpd-bug-report-{}'.format(datetime.now().strftime('%Y%m%d'))
+    component = os.path.join(os.path.expanduser('~'), filename)
+
+    i = 0
+    while os.path.isfile('{}{}.tar'.format(component, '' if not i else '-{}'.format(i))):
+        i += 1
+
+    return '{}{}.tar'.format(component, '' if not i else '-{}'.format(i))
+
+
+def create_bugreport_tar(full_tar_name: str,
+                         log_path: Optional[str]='',
+                         full_config_file: Optional[str]='') -> bool:
+    """
+    Create a tar file containing log and configuration files.
+
+    If the file already exists, do nothing.
+
+    :param full_tar_name: the full path in which to create the tar file
+    :param log_path: path to the log files
+    :param full_config_file: the full path and file of the configuration file
+    :return: True if tar file created, else False
+    """
+
+    if os.path.isfile(full_tar_name):
+        logging.error("Cannot create bug report tarfile, because it already exists")
+        return False
+
+    if not log_path:
+        log_path = os.path.join(xdg.BaseDirectory.xdg_cache_home, 'rapid-photo-downloader', 'log')
+
+    if not full_config_file:
+        config_dir = os.path.join(xdg.BaseDirectory.xdg_config_home, 'Rapid Photo Downloader')
+        config_file = 'Rapid Photo Downloader.conf'
+    else:
+        config_dir, config_file = os.path.split(full_config_file)
+
+    curr_dir = os.getcwd()
+    created = False
+
+    try:
+        with tarfile.open(full_tar_name, 'x') as t:
+            os.chdir(log_path)
+            for l in glob('*'):
+                t.add(l)
+            os.chdir(config_dir)
+            t.add(config_file)
+    except FileNotFoundError as e:
+        logging.error(
+            "When creating a bug report tar file, the directory or file %s does not exist", e.filename
+        )
+    except Exception:
+        logging.exception("Unexpected error when creating bug report tar file")
+    else:
+        created = True
+
+    try:
+        os.chdir(curr_dir)
+    except FileNotFoundError:
+        pass
+
+    return created
 
 
 def current_version_is_dev_version(current_version=None) -> bool: